Create make_small_cat.py
This script randomly selects a set of LRG pairs out of the whole catalog and generates two ASCII files:  one containing all of the information about the pairs, and one containing the locations, redshifts, and LRG pair IDs of all objects neighboring one or more of these pairs.  The purpose of this script is to create smaller catalogs which are representative of the entire set of pairs, so that testing and debugging of code can be done on a local machine rather than on a cluster.