Create gen_mock_cats
Generates two catalogs, one containing the necessary information for a single mock LRG pair and one containing the necessary information for the mock LRG pair's neighboring objects.  As the script is currently set up, the neighbors just form one concentric circle at each radial bin, with all circles centered on the midpoint of the LRG pair.  This is not set up to use proper spherical geometry at present, and so the midpoint should be set to have a dec near zero so that using RA and dec as x and y on a cartesian surface is a valid approximation.